reorganize tasks
